==> scripts/0-trait_dataset_exploration.Rmd <==
---
title: "Exploring the trait dataset file"
author: "Patrick Pata"
date: '2023-01-09'
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Load libraries and files
```{r}
library(here)
source(here("functions/R_package_check.R"))
prepareLibrary()
source(here("functions/toolkit.R"))

theme_set(theme_bw())
set_here(path = "..")

# Load trait dataset metadata tables
directory.trait <- openxlsx::read.xlsx(here("tables/trait_directory_20230207.xlsx"))
directory.lifestage <- openxlsx::read.xlsx(here("tables/lifestage_directory_20230207.xlsx"))
taxonomy <- openxlsx::read.xlsx(here("tables/taxonomy_table_20230316.xlsx"))
dataset.format <- openxlsx::read.xlsx(here("tables/trait_dataset_standard_format_20230403.xlsx"))[-1,]

# Load the trait dataset
# # Load Level 1 dataset
# load("data/Trait_dataset_level1/trait_dataset_level1-2023-04-06.RData")

# Load Level 2 dataset
load("data/Trait_dataset_level2/trait_dataset_level2-2023-04-06.RData")
```


# Contents of the trait dataset
The trait dataset is provided in two levels. Level 1 (traits.lvl1) contains individual-level records which can have multiple measurements of a trait for a species. Level 2 (traits.lvl2) contains species-level averages for numerical traits. Only taxa at the species or subspecies ranks are provided in the level 2 data. For the exploration below, we will look into the level 2 data.
```{r}
# Assign the trait data level to analyze
zooplankton.traits <- traits.lvl2 

# Structure of the dataset
# The dataset starts with identifiers (cols 1-5), followed by the core trait information of the scientific name, trait name, trait value, trait unit, and data type. Many traits have ancillary information (cols 11-26) such as life stage, temperature, and notes. This are followed by the references (cols 27-30) and the taxonomic information (31-41) associated with the trait record. All taxa have a taxonID and most have an aphiaID to link it with the WoRMS database.Details regarding the data provenance of the original trait record are retained in columns 42-54. 
colnames(zooplankton.traits)

# What traits are available?
# The current dataset has 36 traits distributed to 81 traitNames and assigned to 5 general trait buckets. Traits have a dataset-specific traitName which were assigned with standardized trait units. A single trait may be represented in multiple ways and thus have unique traitIDs and traitNames (e.g., the nitrogen content trait can be "nitrogenTotal" which is the bulk total nitrogen of an individual in mg or "nitrogenPDW" which is the percent of nitrogen relative to the dry weight). 
length(unique(directory.trait$traitName))
length(unique(directory.trait$trait))
unique(directory.trait$traitBucket)
unique(directory.trait$dataType)

# Table of traits
kable(select(directory.trait, traitBucket, trait, traitName, traitUnit, dataType))


# The trait dataset contains information in 3 data types. "Continuous" are numerical traits while categorical traits are coded as characters. Categorical traits may also be represented as binary records which can be filtered using dataType == "binary". For binary records, 1 represents presence of a trait and 0 represents absence.
unique(directory.trait$dataType)

# For now, we can review the coverage of traits without the binary traits.
zooplankton.traits %>%
  filter(dataType == "binary") %>% 
  distinct(traitName)
```

# Trait coverage
## Species-TraitName
This lists the traits and counts the number of species-level records for a specific traitName.
```{r, fig.height=3}
# This is the number of taxon observation per trait. The number of taxa with trait records vary with bodyLengthMax having the most trait data. The coverage of rate trait data is limited because this dataset includes only trait observations with temperature information.
traits.summary <- zooplankton.traits %>% 
  filter(dataType != "binary") %>% 
  # Add information about the trait bucket 
  left_join(distinct(directory.trait, traitBucket, traitID),
            by = "traitID") %>% 
  # Assign trait buckets as factors
  mutate(traitBucket = factor(traitBucket, 
                              levels = c("morphological","composition",
                                         "physiological","behavioral",
                                         "life history"))) %>% 
  # Isolate relevant columns
  distinct(traitBucket, traitID, traitName, taxonID) %>% 
  # Add up the number of taxon records per trait
  group_by(traitBucket, traitName, traitID) %>% 
  summarise(Nspecies = n()) %>% 
  ungroup() %>% 
  # Arrange the traits in decreasing order for the figure
  arrange(-Nspecies) %>% 
  mutate(traitName = fct_reorder(traitName, Nspecies))
  

# Stacked barplot with count at end
trait.bar <- ggplot(traits.summary, aes(x = traitName, y = Nspecies)) +
  geom_bar(stat = "identity") +
  # Set axis break
  scale_y_continuous(breaks = c(0, 250, 500, 750, 1000, 3000),
                     limits = c(0, 3180)) +
  scale_y_break(c(1250,2980), expand = FALSE) +
  # Add number of species at the end of the bar
  geom_text(aes(label = after_stat(y), group = traitName),
            stat = "summary", fun = sum, hjust = -0.1,
            size = 3) +
  coord_flip() +
  xlab("Trait") + ylab("# Species") +
  # facets proportioned by number of traits
  facet_grid(traitBucket ~., scales = "free_y", space = "free", switch = "y") +
  theme(strip.placement = "outside") 
trait.bar
```
## Species-TraitName by major group
We can visualize the taxonomic coverage by major group.
```{r, fig.height=3}
# Assign colors to major groups
majorgroup.colors <- data.frame(
  majorgroup = c("Calanoid","Non-calanoid","Amphipod","Decapod","Euphausiid",
                 "Mysid","Ostracod","Polychaete","Pteropod",
                 "Chaetognath","Appendicularian","Thaliacean","Cladoceran",
                 "Hydromedusae","Siphonophore","Scyphomedusae","Ctenophore"),
  color = c("blue","yellowgreen","cornflowerblue","lightseagreen","forestgreen",
            "darkseagreen2","cyan3","tan1","olivedrab",
            "darkorchid","violetred","hotpink3","saddlebrown",
            "tomato1","goldenrod3","yellow3","red4"))
            
# This is similar to the chunk above but with different groupings
traits.summary <- zooplankton.traits %>% 
  filter(dataType != "binary") %>% 
  # Add information about the trait bucket 
  left_join(distinct(directory.trait, traitBucket, traitID),
            by = "traitID") %>% 
  # Assign trait buckets as factors
  mutate(traitBucket = factor(traitBucket, 
                              levels = c("morphological","composition",
                                         "physiological","behavioral",
                                         "life history"))) %>% 
  # Isolate relevant columns
  distinct(traitBucket, traitID, traitName, majorgroup, taxonID) %>% 
  # Add up the number of taxon records per trait
  group_by(traitBucket, traitName, majorgroup, traitID) %>% 
  # Set the levels of the major groups
  mutate(majorgroup = factor(majorgroup, 
                             levels = majorgroup.colors$majorgroup)) %>% 
  summarise(Nspecies = n()) %>% 
  ungroup() %>% 
  # Arrange the traits in decreasing order for the figure
  group_by(traitBucket, traitName) %>% 
  mutate(Nspecies.total = sum(Nspecies)) %>% 
  ungroup() %>% 
  arrange(-Nspecies.total) %>% 
  mutate(traitName = fct_reorder(traitName, Nspecies.total)) 
  
# Stacked barplot with count at end
trait.bar <- ggplot(traits.summary, aes(x = traitName, y = Nspecies,
                                        fill = majorgroup)) +
  geom_bar(stat = "identity") +
  # Set axis break
  scale_y_continuous(breaks = c(0, 250, 500, 750, 1000, 3000),
                     limits = c(0, 3180)) +
  scale_y_break(c(1250,2980), expand = FALSE) +
  # Add number of species at the end of the bar
  geom_text(aes(label = after_stat(y), group = traitName),
            stat = "summary", fun = sum, hjust = -0.1,
            size = 3) +
  scale_fill_manual(values = majorgroup.colors$color,
                    limits = majorgroup.colors$majorgroup,
                    name = "Major group") +
  coord_flip() +
  xlab("Trait") + ylab("# Species") +
  # facets proportioned by number of traits
  facet_grid(traitBucket ~., scales = "free_y", space = "free", switch = "y") +
  theme(strip.placement = "outside") 
trait.bar
```

## Number of traits per species
Here, we inspect the number of traits per species and continue to check how evenly distributed the dataset is. This figure shows that most species have only a few traits assigned to them, but there are a few species with a lot of trait information known.
```{r}
traits.summary.2 <- zooplankton.traits %>% 
  filter(dataType != "binary") %>% 
  distinct(traitName, scientificName, taxonID) %>% 
  group_by(taxonID, scientificName) %>% 
  summarise(Ntraits = n()) %>% 
  arrange(-Ntraits)

head(traits.summary.2)

traits.summary.2 <- traits.summary.2 %>% 
  group_by(Ntraits) %>% 
  summarise(Ntraits.total = sum(Ntraits)) %>% 
  ungroup() %>% 
  arrange(-Ntraits)

ggplot(traits.summary.2, aes(x = Ntraits, y = Ntraits.total)) +
  geom_point(size = 2, colour = "black") +
  geom_segment( aes(x=Ntraits, xend=Ntraits, y=0, yend=Ntraits.total)) +
  xlab("Number of traits per species") +
  ylab("Number of records")
```

# Inspect specific traits

## Numerical traits: average and range of trait between major groups
```{r}
# Numerical traits
# As an example, we can inspect the range of body lengths in each major group.
body.length <- zooplankton.traits %>% 
  filter(traitName == "bodyLengthMax") %>% 
  mutate(traitValue = as.numeric(traitValue)) %>% 
  # calculate the mean to sort major groups
  group_by(majorgroup) %>% 
  mutate(mean.size = mean(traitValue)) %>% 
  ungroup() %>% 
  mutate(majorgroup = fct_reorder(majorgroup, mean.size)) 
  
# In this dataset, body length varies across four orders of magnitude. Many major groups have overlaping and similar body sizes. This suggests that if we approach research questions from, say, a size-spectrum perspective, taxonomic differences between major groups might not be that important.
ggplot(body.length, aes(x = majorgroup, y = traitValue)) +
  geom_violin() +
  # visualize the mean
  stat_summary(fun.data = "mean_cl_boot", geom = "pointrange",
               colour = "red") +
  scale_y_continuous(labels = scaleFUN, trans = "log10") +
  theme(axis.text.x = element_text(angle = 45,  hjust=1))

# Although, there would be taxonomic nuances in how size is measured. A measurement of total body length for individuals with an elongated body plan is typical but for more spherical and/or colonial organisms, length could be based on a specific body axis or individual in colonial organisms. Below lists the various body length axes in the dataset.
(cleanStrings(unique(body.length$sizeType)))
  
# One important trait is nitrogen content. For example, this can be used a currency in biogeochemical models.  When considering composition traits, it is important to distinguish total or percent-weight trait values and choose which would be most appropirate for a particular analysis.
nitrogen.content <- zooplankton.traits %>% 
  # inspect total nitrogen content and the fraction of dry weight which is nitrogen
  filter(traitName %in% c("nitrogenTotal", "nitrogenPDW")) %>% 
  mutate(traitValue = as.numeric(traitValue))
  
ggplot(nitrogen.content, aes(x = majorgroup, y = traitValue)) +
  geom_violin() +
  # visualize the mean
  stat_summary(fun.data = "mean_cl_boot", geom = "pointrange",
               colour = "red") +
  scale_y_continuous(labels = scaleFUN, trans = "log10") +
  theme(axis.text.x = element_text(angle = 45,  hjust=1)) +
  facet_wrap(~traitName, ncol = 1, scales = "free_y")
```


## Categorical traits: inspect proportions
In this dataset, categorical traits, particularly behavioral traits, are some of the most common records. Although numerical traits are useful (and much needed) in quantitative models, categorical traits are often used to characterize zooplankton assemblages or even as predictor variables. Here we inspect some of the commonly utilized categorical traits in zooplankton studies: reproduction mode, feeding mode, trophic group, and vertical distribution. 

Note that in all these traits, an organism can be associated with multiple trait values (e.g., salps have a sexual broadcasting reproduction strategy and an asexual reproduction strategy). This should be considered when performing analysis using categorical traits. For now we will visualize all instances of a categorical trait.
```{r}
# Visualize the number of species associated with a particular categorical trait
vertical.distribution <- zooplankton.traits %>% 
  filter(traitName == "verticalDistribution") %>% 
  group_by(traitValue) %>% 
  summarise(Nspecies = n())

ggplot(vertical.distribution, aes(x = traitValue, y = Nspecies)) +
  geom_bar(stat = "identity") +
  coord_flip() +
  xlab("Vertical distribution") + ylab("# Species") +
  scale_x_discrete(limits = rev(c("epipelagic","epimesopelagic","epibathypelagic",
                              "mesopelagic","mesobathypelagic","bathypelagic")))

# The trait information could be accessed by calling the categorical version of the trait (e.g., "reproductionMode"), but for some instances, using the binary version would be easier in analysis or visuals.
repro.mode <- zooplankton.traits %>% 
  filter(traitName %in% c("RM.broadcasting","RM.brooding","RM.asexual"),
         traitValue == 1) %>% 
  # rename trait levels by removing "RM." prefix
  mutate(traitName = str_replace(traitName, "RM.","")) 

# Bar plot of total species with a particular trait value, organized by major group.
repro.mode <- repro.mode %>% 
  group_by(traitName, majorgroup) %>% 
  summarise(Nspecies = n()) 

ggplot(repro.mode, aes(x = traitName, y = Nspecies, 
                       fill = majorgroup)) +
   geom_bar(stat = "identity") +
  scale_fill_manual(values = majorgroup.colors$color,
                    limits = majorgroup.colors$majorgroup,
                    name = "Major group") +
  coord_flip() +
  xlab("Reproduction mode") + ylab("# Species") 

# Proportion of reproduction modes per major group
ggplot(repro.mode, aes(x = majorgroup, y = Nspecies, 
                       fill = traitName)) +
  geom_bar(position = "fill", stat = "identity") +
  coord_flip() +
  scale_y_continuous(labels = scales::percent)  + 
  scale_x_discrete(limits = rev) +
  xlab("Major group") + ylab("% Species") 


# Relationship between trophic group and feeding mode

# There are different ways in which the feeding mode trait is coded. 
unique(filter(zooplankton.traits, grepl("FM.",traitName))$traitName)

# For now, we can consider, ambush, cruise, and current feeding and check if there might be a relationship between feeding mode and the general trophic group a species is often clustered into.
feeding.trophic <- zooplankton.traits %>% 
  filter(traitName %in% c("FM.ambush","FM.cruise","FM.current","FM.particle.feeder"),
         traitValue == 1) %>% 
  # rename the trait variable
  rename(feeding.mode = traitName) %>% 
  # rename trait levels by removing "RM." prefix
  mutate(feeding.mode = str_replace(feeding.mode, "FM.","")) %>% 
  # merge with the column on trophic groups
  left_join(select(filter(zooplankton.traits,
                          traitName %in% c("TG.omnivore","TG.herbivore",
                                           "TG.carnivore","TG.detritivore")),
                   taxonID, trophic.group = traitName),
            by = "taxonID") %>% 
  mutate(trophic.group = str_replace(trophic.group, "TG.","")) %>% 
  relocate(feeding.mode, trophic.group) %>% 
  # only consider instances when information for both traits are available
  filter(!is.na(feeding.mode) & !is.na(trophic.group)) %>% 
  group_by(feeding.mode, trophic.group) %>% 
  summarise(Nspecies = n())

ggplot(feeding.trophic, aes(x = trophic.group, y = Nspecies,
                            fill = feeding.mode)) + 
  geom_bar(position = "fill", stat = "identity") +
  scale_y_continuous(labels = scales::percent)  + 
  xlab("Major group") + ylab("% Species") 

```


## Size as a predictor of trait distributions
Size is commonly referred to as a "master trait" because many biological attributes scales with size. There are multiple measures of size such as length or biomass. For zooplankton, this is is an important concern when considering both crustaceans and gelatinous or soft-bodied species. Gelatinous zooplankton have an "inflated" size because they have a higher water content.
Note that for the figures below, we are simply using the geom_smooth() function of ggplot to estimate and visualize a linear model. If we to formally derive a regression, care must be taken in choosing the regression model and data curation.
```{r}
unique(zooplankton.traits$traitName)

# Because body length is a commonly reported size measurement, we can inspect how the measures of weight scale with length.
size.traits <- zooplankton.traits %>% 
  # filter species with weight measurements
  filter(traitName %in% c("carbonWeight", "dryWeight", "wetWeight")) %>% 
  # Append body length as a column for comparison
  left_join(dplyr::select(filter(zooplankton.traits, 
                                 traitName == "bodyLengthMax"),
                          taxonID, bodyLength = traitValue),
            by = "taxonID") %>% 
  # Filter out instances when there are no length information.
  filter(!is.na(bodyLength)) %>% 
  # Make sure to convert continuous traits to numeric
  mutate(traitValue = as.numeric(traitValue),
         bodyLength = as.numeric(bodyLength))
  
# It is clear that weight scales with length, but there seems to be quite some variability in this relationship.
size.plot <- ggplot(data = size.traits, 
                    aes(x = bodyLength, y = traitValue, color = traitName)) +
  geom_point() +
  # Add labels for inspecting points
  geom_point(aes(text = paste0(scientificName," (",majorgroup,")"))) +
  # Add a linear model
  geom_smooth(method = "lm") +
  xlab("Length (mm)") + ylab("Weight (mg)") +
  # Log transform axis scales
  scale_x_continuous(labels = scaleFUN, trans = "log10") +
  scale_y_continuous(labels = scaleFUN, trans = "log10") 
ggplotly(size.plot)


# To further inspect this scaling, we can separate hard-bodied crustacean zooplankton from gelatinous and soft-bodied zooplankton. 
size.traits <- size.traits %>% 
  # Assign a body plan grouping variable
  mutate(bodyPlan = if_else(phylum == "Arthropoda","hard-body","soft-body"))

# Focus on carbon weight. Here we see that body plan mediates the slope of the weight - length relationship.
size.plot.2 <- ggplot(data = filter(size.traits, traitName == "carbonWeight"), 
                      aes(x = bodyLength, y = traitValue, color = bodyPlan)) +
  # Add labels for inspecting points
  geom_point(aes(text = paste0(scientificName," (",majorgroup,")"))) +
  # Add a linear model
  geom_smooth(method = "lm") +
  xlab("Length (mm)") + ylab("Weight (mg)") +
  # Log transform axis scales
  scale_x_continuous(labels = scaleFUN, trans = "log10") +
  scale_y_continuous(labels = scaleFUN, trans = "log10") 
ggplotly(size.plot.2)


# Allometric scaling of physiological rates with size
# A common application of size as a master trait is using it as a predictor for other traits, especially physiological rates. Here, we inspect respiration rate, which is the most common rate trait in the dataset. Carbon weight will be used as the measure for size.
rate.size <- zooplankton.traits %>% 
  # select respiration rate traits
  filter(traitName %in% c("respirationRate_15C", "respirationRate_WSC_15C")) %>%
  # Append carbon weight as a column for comparison
  left_join(dplyr::select(filter(zooplankton.traits, 
                                 traitName == "carbonWeight"),
                          taxonID, carbonWeight = traitValue),
            by = "taxonID") %>% 
  # Filter out instances when there are no carbon weight information.
  filter(!is.na(carbonWeight)) %>% 
  # Make sure to convert continuous traits to numeric
  mutate(traitValue = as.numeric(traitValue),
         carbonWeight = as.numeric(carbonWeight))

# Here we see carbon weight scales with size. Weight-specific rate values often have flatter slope, and sometimes negative slope. 
rate.size.plot <- ggplot(data = rate.size, 
                      aes(x = carbonWeight, y = traitValue, color = traitName)) +
  # Add labels for inspecting points
  geom_point(aes(text = paste0(scientificName," (",majorgroup,")"))) +
  # Add a linear model
  geom_smooth(method = "lm") +
  # xlab("Length (mm)") + ylab("Weight (mg)") +
  # Log transform axis scales
  scale_x_continuous(labels = scaleFUN, trans = "log10") +
  scale_y_continuous(labels = scaleFUN, trans = "log10") 
ggplotly(rate.size.plot)


# Now, can inspect the size scaling of the excretion rate of N as ammonia. The data source for excretion rates used dry weight as the associated size measure, so we reference excretion rate relative to dry weight.
rate.size.2 <- zooplankton.traits %>% 
  # select respiration rate traits
  filter(traitName %in% c("excretionRateN_15C", "excretionRateN_WSDW_15C")) %>%
  # Append dry weight as a column for comparison
  left_join(dplyr::select(filter(zooplankton.traits, 
                                 traitName == "dryWeight"),
                          taxonID, dryWeight = traitValue),
            by = "taxonID") %>% 
  # Filter out instances when there are no dry weight information.
  filter(!is.na(dryWeight)) %>% 
  # Make sure to convert continuous traits to numeric
  mutate(traitValue = as.numeric(traitValue),
         dryWeight = as.numeric(dryWeight))


rate.size.plot.2 <- ggplot(data = rate.size.2, 
                      aes(x = dryWeight, y = traitValue, color = traitName)) +
  # Add labels for inspecting points
  geom_point(aes(text = paste0(scientificName," (",majorgroup,")"))) +
  # Add a linear model
  geom_smooth(method = "lm") +
  # xlab("Length (mm)") + ylab("Weight (mg)") +
  # Log transform axis scales
  scale_x_continuous(labels = scaleFUN, trans = "log10") +
  scale_y_continuous(labels = scaleFUN, trans = "log10") 
ggplotly(rate.size.plot.2)


```

## Relationship between traits
We can further look into how a trait might affect the distribution of other traits or the allometric scaling relationship of traits.
```{r}
rate.size <- zooplankton.traits %>% 
  # select respiration rate traits
  filter(traitName %in% c("respirationRate_15C", "respirationRate_WSC_15C")) %>%
  # Append carbon weight as a column for comparison
  left_join(dplyr::select(filter(zooplankton.traits, 
                                 traitName == "carbonWeight"),
                          taxonID, carbonWeight = traitValue),
            by = "taxonID") %>% 
  # Filter out instances when there are no carbon weight information.
  filter(!is.na(carbonWeight)) %>% 
  # Make sure to convert continuous traits to numeric
  mutate(traitValue = as.numeric(traitValue),
         carbonWeight = as.numeric(carbonWeight))

# One broad way to classify zooplankton is whether they have a passive or active feeding mode. Passive feeding include ambush feeding behavior while active feeding includes raptorial cruise feeding and current or filter-feeding.
fm.activity <- zooplankton.traits %>% 
  filter(traitName %in% c("FM.passive","FM.active"),
         traitValue == 1) %>% 
  # transform for merging with the size data
  select(taxonID, feeding.activity = traitName) %>% 
  # Check if there are multiple records for the same taxon. This may not be a concern because some species employ multiple behavioral strategies. Note that this may result in duplicated rows when doing a left join. 
  group_by(taxonID) %>% 
  mutate(nvals = n()) %>% 
  arrange(-nvals)

# We know that respiration rate scales with size, but does feeding activity mediate this relationship? To explore this question, we include a feeding activity column to the rate-size figure. Here we see that active feeders have a slightly higher respiration rate compared to passive feeders. But is this a significant difference? Could the taxonomic coverage of the species included in this analysis limit what we can infer? Would it be more informative if the individual-level trait data was used?

rate.size.activity <- rate.size %>% 
  left_join(fm.activity, by = "taxonID") %>% 
  relocate(feeding.activity) %>% 
  filter(!is.na(feeding.activity))

rate.activity.plot <- ggplot(data = filter(rate.size.activity, 
                                           traitName == "respirationRate_15C"),
                             aes(x = carbonWeight, y = traitValue, color = feeding.activity)) +
  # Add labels for inspecting points
  geom_point(aes(text = paste0(scientificName," (",majorgroup,")"))) +
  # Add a linear model
  geom_smooth(method = "lm") +
  ylab("Respiration rate") +
  # Log transform axis scales
  scale_x_continuous(labels = scaleFUN, trans = "log10") +
  scale_y_continuous(labels = scaleFUN, trans = "log10") 
ggplotly(rate.activity.plot)


```
