==> scripts/1-zooplankton_dataset_curation.Rmd <==
---
title: "Curating the zooplankton dataset file"
author: "Patrick Pata"
date: '2023-01-09'
output: html_document
---

In this file, the Strait of Georgia zooplankton dataset is curated and file which summarizes the species-level abundance and biomass is exported. A species list to match the scientific names of the SOG distribution data with the trait database is also exported.

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Load libraries and files

```{r}
library(here)
source(here("functions/R_package_check.R"))
prepareLibrary()
source(here("functions/toolkit.R"))

theme_set(theme_bw())
set_here(path = "..")

# Taxonomy list for matching with trait dataset
taxonomy <- read.csv(here("tables/taxonomy_table_20230628.csv"))
```

# Load and curate the zooplankton net data.

Data source for zooplankton abundance and biomass data: <https://open.canada.ca/data/en/dataset/2822c11d-6b6c-437e-ad65-85f584522adc> Data dictionary: <https://pacgis01.dfo-mpo.gc.ca/FGPPublic/Zooplankton_Data_SoG/Data_dictionary_EN_FR_SoG_zooplankton_data_1996-2018.htm>

The taxonomic resolution of the dataset is at species-stage or size class. See the data dictionary for details. The plankton data was provided as abundance (number of individuals per volume of water) and using the provided value of the average biomass per individual (specific to a species and stage or size class), the biomass is calculated.

View data variables: (For details, see: <https://pacgis01.dfo-mpo.gc.ca/FGPPublic/Zooplankton_Data_SoG/Data_dictionary_EN_FR_SoG_zooplankton_data_1996-2018.htm>)

## Load the zooplankton abundance dataset from net observations
```{r}
zoop.SOG <- read.csv(here("data/Zoopl_SofG_1996-2018_EN.csv"),
                            check.names = F) %>% 
  # Rename some columns
  rename("Longitude" = "Longitude(\xb0N)", "Latitude" = "Latitude(\xb0W)",
         "Mesh_size(um)" = "Mesh_size(\xb5m)",
         "Abundance(indiv/m3)" = "Abundance(m3)") %>% 
  
  # Create new scientific name column.
  mutate(scientificName = paste(Genus,Species)) %>% 
  mutate(scientificName = cleanScientificName(scientificName)) %>% 
  relocate(scientificName) %>% 
  
  # Get stage information
  mutate(Stage = str_remove(Size_class_or_life_stage, scientificName)) %>% 
  mutate(Stage = str_replace_all(Stage, "[[:punct:]]", "")) %>% 
  mutate(Stage = str_remove(Stage, "sp")) %>% 
  mutate(Stage = str_squish(str_trim(Stage))) %>% 
  
  # Calculate biomass as number of individuals per sample * average biomass per individual
  mutate(`Biomass(mg/m3)` = `Abundance(indiv/m3)` * `Average_biomass_per_individual(mg)`) %>% 
  
  # Assign seasons: The spring bloom occurs around March in the Strait of Georgia and so "spring" is set to be March-April-May
  mutate(Season = "Winter") %>% 
  mutate(Season = if_else(Month >= 3 & Month <=5, "Spring", Season),
         Season = if_else(Month >= 6 & Month <=8, "Summer", Season),
         Season = if_else(Month >= 9 & Month <=11, "Fall", Season)) %>% 
  
  # Assign date-time using lubridate
  mutate(datetime = lubridate::make_datetime(Year, Month, Day)) %>% 
  mutate(dayofyear = lubridate::yday(datetime)) %>% 
  
  # Need to include time of day as well     
                                             # as.numeric(unlist(strsplit(Time,split=":"))[1]),
                                             # as.numeric(unlist(strsplit(Time,split=":"))[2])))
  #  tz = "Canada/Pacific")) # setting the time zone now returns an error. Need to figure out why.
  
  # Remove a zero at the 8th character to match with the old index names
  mutate(NetKey = paste0(str_sub(Index,1,7),str_sub(Index,9,)))
```

## Curate the taxonomic extent of the dataset and create a species list
```{r}
zoop.SOG <- zoop.SOG %>% 
  # Exclude microzooplankton
  filter(Phylum %notin% c("Pyrrophycophyta (Division)","Ciliophora",
                            "Alveolates","Unclassified","Protozoa")) %>% 
  
  # Remove select groups which have trait data
  filter(Phylum %in% c("Annelida","Arthropoda","Chaetognatha","Chordata",
                       "Cnidaria","Mollusca","Ctenophora")) %>% 
  
  # Exclude class Scyphozoa because they are too large for a representative sampling of biomass
  # Also exclude larval fish and invertebrates which are not in the trait database
  filter(Class %notin% c("Scyphozoa", "Actinopterygii", "Cephalopoda",
                         "Bivalvia", "Cirripedia", "Ascidiacea", "Anthozoa",
                         "Insecta", "Arachnida","Branchiopoda")) %>% 
  # Exclude non-pelagic polychaetes and orders not in trait database
  filter(Order %notin% c("Canalipalpata","Polychaeta","Isopoda","Lophogastrida",
                         "Cumacea","Decapoda","Neotaenioglossa","Siphonostomatoida",
                         # Exclude records not at genus or species level
                         "Chaetognatha","Ostracoda","Ctenophora")) %>% 
  filter(Genus %notin% c("Phyllodocida","Gammaridea","Oedicerotidae",
                         "Phoxocephalidae","Haustoriidae","Euphausiacea",
                         "Euchaetidae","Aetideidae","Cyclopoida",
                         "Harpacticoida","Copepoda","Monstrilloida",
                         "Oncaeidae","Conchoecinae","Philomedidae",
                         "Anthomedusae","Hydrozoa","Medusae","Mitrocomidae",
                         "Pandeidae","Siphonophora","Narcomedusae",
                         "Rhopalonematidae","Gastropoda","Syllidae",
                         "Lysianassidae","Euphausiidae","Nereidae",
                         "Macrochiron","Gastropteron","Aeginidae","Polynoidae",
                         "Calanoida")) %>% 
  filter(Species %notin% c("marshallae+pacificus")) %>% 
  # No trait info for these copepods
  filter(scientificName %notin% c("Atrophia minuta","Lubbockia wilsonae")) %>% 
  
  # Depth curation: >= 100 m bottom depth and tows >= 70% of bottom depth (following Perry et al. 2021)
  mutate(Tow.Bottom = (as.numeric(`Tow_start_depth(m)`)/as.numeric(`Bottom_depth(m)`))) %>% 
  filter(`Bottom_depth(m)` >= 100 & Tow.Bottom >= 0.7) %>% 
  droplevels() %>% 
  
  # Correct the adult euphausiid biomass and abundance to account for reduced catches during daylight due to net avoidance (Mackas et al. 2013).
  mutate(`Abundance(indiv/m3)` = if_else(Order == "Euphausiacea" & 
                                           Stage %in% c("s3","F","M") & 
                                           Twilight=="Daylight",
                                         `Abundance(indiv/m3)` * 3, 
                                         `Abundance(indiv/m3)`)) %>% 
  mutate(`Biomass(mg/m3)` = if_else(Order == "Euphausiacea" & 
                                           Stage %in% c("s3","F","M") & 
                                           Twilight=="Daylight",
                                         `Biomass(mg/m3)` * 3, 
                                         `Biomass(mg/m3)`)) %>% 
  
  # Exclude a sample with unusually low species count
  filter(NetKey != "UBC999999000301") %>% 
  
  # Rename 6 taxa to match database names and accepted names in WORMS
  mutate(scientificName = if_else(scientificName == "Alacia pseudoalata",
                                  "Alacia major", scientificName),
         scientificName = if_else(scientificName == "Rhynchonereella angelina",
                                  "Rhynchonereella angelini", scientificName),
         scientificName = if_else(scientificName == "Conchoecia rudjakovi",
                                  "Conchoecia rudyakovi", scientificName),
         scientificName = if_else(scientificName == "Gaidius",
                                  "Gaetanus", scientificName),
         Genus = if_else(Genus == "Gaidius",
                                  "Gaetanus", Genus),
         scientificName = if_else(scientificName == "Mesosagitta decipiens",
                                  "Decipisagitta decipiens", scientificName),
         scientificName = if_else(scientificName == "Meterythrops robusta",
                                  "Meterythrops robustus", scientificName) )	%>% 

  # Convert some taxa identified to genus level to the only species found in the area. This is equivalent to generalizing to the genus level but makes it easier to match with trait data.
  mutate(scientificName = if_else(scientificName == "Themisto",
                                  "Themisto pacifica", scientificName),
         scientificName = if_else(scientificName == "Fritillaria",
                                  "Fritillaria borealis", scientificName),
         scientificName = if_else(scientificName == "Clione",
                                  "Clione limacina", scientificName),
         scientificName = if_else(scientificName == "Centropages",
                                  "Centropages abdominalis", scientificName),
         scientificName = if_else(scientificName == "Eurytemora",
                                  "Eurytemora americana", scientificName),
         scientificName = if_else(scientificName == "Lubbockia",
                                  "Lubbockia wilsonae", scientificName),
         scientificName = if_else(scientificName == "Bradyidius",
                                  "Bradyidius saanichi", scientificName),
         scientificName = if_else(scientificName == "Hyperia",
                                  "Hyperia medusarum", scientificName),
         scientificName = if_else(scientificName == "Scolecithricella",
                                  "Scolecithricella minor", scientificName),
         scientificName = if_else(scientificName == "Euchirella",
                                  "Euchirella pseudopulchra", scientificName),
         scientificName = if_else(scientificName == "Pleurobrachia",
                                  "Pleurobrachia bachei", scientificName),
         scientificName = if_else(scientificName == "Bougainvillia",
                                  "Bougainvillia multitentaculata", scientificName),
         scientificName = if_else(scientificName == "Primno",
                                  "Primno abyssalis", scientificName),
         scientificName = if_else(scientificName == "Tortanus",
                                  "Tortanus discaudatus", scientificName),
         scientificName = if_else(scientificName == "Euphysa",
                                  "Euphysa japonica", scientificName),
         scientificName = if_else(scientificName == "Epilabidocera",
                                  "Epilabidocera longipedata", scientificName),
         scientificName = if_else(scientificName == "Paraeuchaeta",
                                  "Paraeuchaeta elongata", scientificName) ) %>% 
  
  # Exclude early life stages: eggs, nauplii, zoea, and copepodites with unclear stages
  filter(Stage %notin% c("eggs s1","nauplii s1","zoea s1","polytroch s0","s0",
                         "13","14")) %>% 

  # Exclude taxa that only appear in 0.5% of the samples of the dataset (very rare taxa)
  mutate(nsamples = length(unique(NetKey))) %>% 
  group_by(scientificName) %>% 
  mutate(nobs = n()) %>% 
  ungroup() %>% 
  filter(nobs/nsamples > 0.005) %>% 
  select(-c(nobs, nsamples))%>%

  # Match the distribution data and trait database species list
  left_join(distinct(taxonomy, scientificName = verbatimScientificName,
                     taxonID, taxonRank,
                     acceptedNameUsage, acceptedNameUsageID, majorgroup),
            by = "scientificName") 
  
# List the species in the curated dataset
sp.list <- zoop.SOG %>% 
  distinct(taxonRank, scientificName, Phylum, Class, Order, Family, Genus,
           taxonID, majorgroup, acceptedNameUsage, acceptedNameUsageID) 
```

# Calculate alpha diversity and add to metadata
```{r}
# Calculate total abundance, biomass, and diversity metrics for each sample
zoop.SOG.alphadiv <- zoop.SOG %>% 
  group_by(Index, scientificName) %>% 
  summarise(`Abundance(indiv/m3)` = sum(`Abundance(indiv/m3)`),
            .groups = "drop") %>% 
  group_by(Index) %>% 
  summarise(Richness = specnumber(`Abundance(indiv/m3)`),
            Diversity.Shannon = diversity(`Abundance(indiv/m3)`, index = "shannon"),
            Diversity.Simpson = diversity(`Abundance(indiv/m3)`, index = "simpson"),
            .groups = "drop") %>% 
  mutate(Evenness = Diversity.Shannon/log(Richness))


# Create and save a metadata file of the curated zooplankton data.
zoop.SOG.meta <- zoop.SOG %>% 
  select(Index, Region, Station, Longitude, Latitude,
         Year, Month, Day, Time, Season, datetime, dayofyear,
         Twilight, Net_type,
         `Net_Diameter(m)`, `Mesh_size(um)`, 
         `Tow_start_depth(m)`,`Tow_end_depth(m)`,
         `Bottom_depth(m)`,`Volume(m3)`,
         NetKey) %>% 
  distinct() %>% 
  left_join(zoop.SOG.alphadiv, by = "Index")
```


# Export the curated species distribution dataset
```{r}
save(zoop.SOG, zoop.SOG.meta, sp.list,
     file = here("data/curated_zoop_distribution_20230801.RData"))

```


# --- Extra/Deprecated scripts --- 
Create a species-level summary of the zooplankton dataset. This removes life stage or size class information so better to curate which life stages to include in an analysis before calculating species averages:

```{r}
zoop.SOG.sp <- zoop.SOG %>% 
  select(-c("Average_biomass_per_individual(mg)","Size_class_or_life_stage")) %>% 
  group_by(scientificName) %>% 
  mutate(`Abundance(indiv/m3)` = sum(`Abundance(indiv/m3)`),
         `Biomass(mg/m3)` = sum(`Biomass(mg/m3)`)) %>% 
  ungroup() %>% 
  distinct()

zoop.SOG.sp
# TODO: extract a column of life stages
```

There are available data from 1996 to 2018 but seasonal sampling frequencies vary per year:

```{r}
sample.summary <- zoop.SOG.meta %>% 
  group_by(Year, Season) %>% 
  count()
kable(sample.summary)
```

Which stations get sampled the most?

```{r}
zoop.SOG.meta %>% 
  group_by(Station) %>% 
  count() %>% 
  arrange(-n)
```

# Visualize the distribution of samples in space and time

Plot map of Strait of Georgia:

```{r}
# PolySet data frames for mapping the Strait of Georgia from the PBSmapping package
data(nepacLL)
data(nepacLLhigh) #higher resolution maps
colnames(nepacLLhigh) <- c("group", "POS", "lon", "lat") # rename columns
colnames(nepacLL) <- c("group", "POS", "lon", "lat")


# This is the high resolution base map for the SoG. Please change data = "nepacLL" if a lower resolution without some small islands is preferred.
SOGmap <- ggplot() + 
  geom_polygon(data=nepacLLhigh, aes(lon, lat, group=group), 
               fill="grey85", size=0.2, color="black") +
  coord_map(projection='mercator', 
            xlim = c(-126, -122), ylim = c(48, 51)) + 
  theme(axis.title = element_blank(), axis.text = element_blank()) +
  theme_classic() +
  # Spatial distribution of samples. Note that there are multiple overlapping points and there are a few stations that are regularly sampled.
  geom_point(data = zoop.SOG.meta, aes(x = Longitude, y = Latitude,
                                       color = `Bottom_depth(m)`)) +
  ggtitle("Distribution of zooplankton net samples")

SOGmap
```

Plot the temporal distribution of samples:

```{r}
ggplot(data = zoop.SOG.meta, aes(x = dayofyear, y = Year)) +
  geom_point(aes(color = Twilight)) +
  # Roughly mark seasons
  geom_vline(xintercept = c(32,120,212,304), linetype="dotted", color = "blue")
```
