==> scripts/3-oceanography_dataset_exploration.Rmd <==
---
title: "Oceanography and satellite data for SOG"
author: "Patrick Pata"
date: '2023-01-10'
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Load libraries and files

```{r}
library(here)
source(here("functions/R_package_check.R"))
prepareLibrary()
source(here("functions/toolkit.R"))
source(here("functions/getClimateIndices.R"))

theme_set(theme_bw())
set_here(path = "..")

# Load net metadata 
load(here("data/curated_zoop_distribution_20230801.RData"))
rm(zoop.SOG, sp.list)

# Load SOG physics data
load( here("data/SOG_physics_data_2023_01_10.RData"))

# Load the satellite data
load(here("data/SOG_satellite_chl_19972022.RData"))

```

# Create sample x environmental variables matrix
```{r}
# Extract global climate indices
climate <- getClimateIndices(index = c("PDO","NPGO","MEI"), 
                             year.s = 1995, year.e = 2020) %>% 
  mutate(Index = toupper(Index)) %>% 
  pivot_wider(names_from = Index, values_from = Value)

# Match environmental data with the net metadata
zoop.SOG.envi <- zoop.SOG.meta %>% 
  select(Index, NetKey, Season, Region, Longitude, Latitude, 
         Year, Month, dayofyear, 
         Depth = `Bottom_depth(m)`) %>% 
  # Add CTD data (for 530 of the samples). There are not enough chemistry data to match each sample. For the CTD data, select the surface (0-10m) and net depth (0-net tow depth) averages temperature, salinity, and density values.
  left_join(select(physics.ctd, NetKey, TempI10, TempINet, SaliI10, SaliINet,
                   DensI10, DensINet), by = "NetKey") %>% 
  # Add climate information
  left_join(climate, by = join_by(Year,Month)) %>% 
  select(-NetKey)


# For now, use average satellite chlorophyll-a values for a year, month, and region and assign these to samples collected during these periods.
# TODO can further refine the specificity of the spatial matching by finding the satellite grid cell a sample was collected from.
chl.timeseries <- satellite.sog %>% 
  # Select data from 1997 to 2018
  filter(year >= 1997 & year <= 2018) %>% 
  # The coverage of grid cells with chlorophyll data varies every month due to cloud cover. For calculating the SOG chlorophyll average, only include grid cells which are often sampled throughout the time series, set to an arbitrary threshold of N data points = 100. 
  group_by(lon, lat) %>% 
  mutate(Ndata = sum(!is.na(chl))) %>% 
  ungroup() %>% 
  filter(Ndata >= 100) %>% 
  # Calculating the average chlorophyll for SOG. Also divide into northern and central SOG roughly at 49.5 degrees latitude.
  # overall average
  group_by(year, month) %>% 
  mutate(chla.sog = mean(chl, na.rm = T),
         chla.sog.sd = sd(chl, na.rm = T)) %>% 
  # regional average
  mutate(Region = if_else(lat >= 49.5, 
                          "Northern Strait of Georgia", 
                          "Central Strait of Georgia")) %>%
  group_by(year, month, Region) %>%
  mutate(chla.sog.region = mean(chl, na.rm = T),
         chla.sog.region.sd = sd(chl, na.rm = T)) %>%
  ungroup() %>% 
  # summarise dataframe
  select(-c(lon,lat,chl,Ndata)) %>% 
  distinct() %>% 
  rename(Year = year, Month = month)


# Merge the CTD and satellite chlorophyll data
zoop.SOG.envi <- zoop.SOG.envi %>% 
  left_join(select(chl.timeseries, Year, Month, Region, Chla = chla.sog.region),
            by = c("Year","Month","Region"))


# Output sites x environment data
save(zoop.SOG.envi,
     file = here("data/temporary_site_environment_20230821.RData"))

# TODO depending on the spatial and temporal resolution of the analysis, the CTD, satellite, Fraser discharge, and lighthouse data can be averaged first before matching with the net data.

```

# Number of net samples per month and region
```{r}
AA <- zoop.SOG.meta %>% 
  group_by(Year, Month) %>% 
  summarise(nsamples = n(), richness.mean = mean(Richness), richness.sd = sd(Richness),
            .groups = "drop") %>% 
  mutate(Year.Month = Year + (Month-1)/12)

hist(AA$nsamples)

ggplot(AA, aes(x = Year.Month, y = nsamples, color = richness.mean)) +
  geom_point(size = 2) +
  scale_color_viridis()

```



# Distribution of CTD samples

A common sensor in oceanographic studies is the CTD (conductivity, temperature, depth). Conductivity is used to calculate for salinity. Additional sensors, such as a dissolved oxygen sensor, are often attached to the CTD system. A CTD sensor is typically deployed from research vessels and drops to a target depth, often just above the bottom of the ocean. It logs physical variables along the way and these are generally interpolated into 1 m depth bins.

The physics dataset includes CTD data for salinity, temperature, and oxygen. Density was calculated from the temperature, salinity, and pressure information. The data are provided as depth-integrated average values from the surface to either 10 m, 50 m, or net tow depth. The variables have a suffix of I10, I50, or INet corresponding to the integration depth.

Here, we will first explore the spatial and temporal distribution of the CTD data. Notice that the CTD data were collected all over the Strait of Georgia and there are may overlapping points. Both spatially and temporally, the sampling frequency is uneven. Although there are permanent stations which are regularly monitored.

Structure of the physics data:

```{r}
colnames(physics.ctd)
```

Plot the distribution of CTD samples:

```{r}
# PolySet data frames for mapping the Strait of Georgia from the PBSmapping package
data(nepacLL)
data(nepacLLhigh) #higher resolution maps
colnames(nepacLLhigh) <- c("group", "POS", "lon", "lat") # rename columns
colnames(nepacLL) <- c("group", "POS", "lon", "lat")

# Subset columns with CTD data
ctd.SOG <- physics.meta %>% 
  filter(!is.na(CTDKey))
  
# This is the high resolution base map for the SoG. Please change data = "nepacLL" if a lower resolution without some small islands is preferred.
SOGmap <- ggplot() + 
  geom_polygon(data=nepacLLhigh, aes(lon, lat, group=group), 
               fill="grey85", size=0.2, color="black") +
  coord_map(projection='mercator', 
            xlim = c(-126, -122), ylim = c(48, 51)) + 
  theme(axis.title = element_blank(), axis.text = element_blank()) +
  theme_classic()

# Spatial distribution of samples. Note that there are multiple overlapping points and there are a few stations that are regularly sampled.
SOGmap +
  geom_point(data = ctd.SOG, aes(x = Longitude, y = Latitude,
                                       color = `Depth_Bottom.m.`),
             alpha = 0.5) +
  ggtitle("Distribution of CTD samples")
```

Temporal distribution of samples

```{r}
# Temporal distribution of samples
ggplot(data = ctd.SOG, aes(x = dayofyear, y = Year)) +
  geom_point(aes(color = Twilight), alpha = 0.5) +
  # Roughly mark seasons
  geom_vline(xintercept = c(32,120,212,304), linetype="dotted", color = "blue")
```

List most common research stations:

```{r}
# List most common research stations
ctd.SOG %>% 
  group_by(Station) %>% 
  count() %>% 
  arrange(-n) %>% 
  head(10)
```

Update the map and temporal distribution to reflect these stations:

```{r}
# Create a new variable Station2 which indicates if in the three main stations of not
ctd.SOG <- ctd.SOG %>% 
  mutate(Station2 = if_else(Station %in% c("GEO1","CPF1","CPF2"),
                            Station, "Other"))
# Map
SOGmap +
  geom_point(data = ctd.SOG, aes(x = Longitude, y = Latitude,
                                       color = Station2),
             alpha = 0.5) +
  ggtitle("Distribution of CTD samples")

# Temporal distribution
ggplot(data = ctd.SOG, aes(x = dayofyear, y = Year)) +
  geom_point(aes(color = Station2), alpha = 0.5, size = 2) +
  # Roughly mark seasons
  geom_vline(xintercept = c(32,120,212,304), linetype="dotted", color = "blue")

```

# Time series of ctd data

Here, we do a simple curation of the physics data and present a monthly averaged CTD time series. Notice that the temporal distribution of sampling is not consistent.

```{r}
ctd.ts <- physics.ctd %>% 
  # Merge some metadata
  left_join(distinct(ctd.SOG, CTDKey, Station, Longitude, Latitude,
                     Year, Month, Day), by = "CTDKey") %>% 
  # Here, we will only consider plotting the time series of the 3 most common stations.
  filter(Station %in% c("GEO1","CPF1","CPF2")) %>% 
  # Lengthen the data frame
  pivot_longer(cols = c("TempI10", "TempI50", "TempINet", 
                        "SaliI10", "SaliI50", "SaliINet",
                        "DensI10", "DensI50", "DensINet",
                        "OxygI10", "OxygI50", "OxygINet"),
               names_to = "Variable", values_to = "Value") %>% 
  filter(!is.na(Value)) %>% 
  # Calculate monthly averages
  group_by(Year, Month, Variable) %>% 
  summarise(Value = mean(Value), n = n()) %>% 
  ungroup() %>% 
  # Derive Year-Month variable for plotting
  mutate(YearMonth = Year + (Month - 1)/12)

# Visualize the time series for the upper 50 m  CTD values
ggplot(data = filter(ctd.ts, Variable %in% c("TempI50","SaliI50","DensI50","OxygI50")),
       aes(YearMonth, Value)) + 
  geom_point() +
  geom_line() +
  facet_wrap(~Variable, ncol = 1, scales = "free_y")

```

# Satellite data

Chlorophyll a and other ocean color variables are collected from satellites. Here, we explore the gridded monthly satellite-derived chlorophyll values for the Strait of Georgia. First, we filter the dataset to a time period and only include grid cells which often have enough data points. Then, the average chlorophyll-a value for each month is calculated. Finally, an anomaly time series is derived.

```{r}
# View distribution of grid cells subset from the global satellite data
sat.grid <- satellite.sog %>% 
  distinct(lon, lat)

SOGmap +
  geom_point(data = distinct(sat.grid, lon, lat), 
             aes(x = lon, y = lat))
# min(satellite.sog$year)
# max(satellite.sog$year)
```

Plot time series:

```{r}
chl.timeseries <- satellite.sog %>% 
  
  # Select data from 1997 to 2018
  filter(year >= 1997 & year <= 2018) %>% 
  
  # The coverage of grid cells with chlorophyll data varies every month due to cloud cover. For calculating the SOG chlorophyll average, only include grid cells which are often sampled throughout the time series, set to an arbitrary threshold of N data points = 100. 
  group_by(lon, lat) %>% 
  mutate(Ndata = sum(!is.na(chl))) %>% 
  ungroup() %>% 
  filter(Ndata >= 100) %>% 
 
  # Calculating the average chlorophyll for SOG. Also divide into northern and central SOG roughly at 49.5 degrees latitude.
  # overall average
  group_by(year, month) %>% 
  mutate(chla.sog = mean(chl, na.rm = T),
         chla.sog.sd = sd(chl, na.rm = T)) %>% 
  # regional average
  mutate(region = if_else(lat >= 49.5, "northern", "central")) %>%
  group_by(year, month, region) %>%
  mutate(chla.sog.region = mean(chl, na.rm = T),
         chla.sog.region.sd = sd(chl, na.rm = T)) %>%
  ungroup() %>% 
  
  # summarise dataframe
  select(-c(lon,lat,chl,Ndata)) %>% 
  distinct() %>% 
  
  # create year.month variable for timeseries
  mutate(year.month = year + (month-1)/12)



# Plot the timeseries
# overall SOG
ggplot(data = distinct(chl.timeseries, year.month, .keep_all = T ),
       aes(x = year.month, y = chla.sog)) +
  geom_point() +
  geom_line() 
# per region
ggplot(data = chl.timeseries,
       aes(x = year.month, y = chla.sog.region,
           color = region)) +
  geom_point() +
  geom_line()

```

Anomally time series: For time series data, it is often interesting to look into the anomaly of a parameter relative to a climatological mean. Here we calculate the monthly climatology of chla in the SOG and then, calculate the anomaly time series.

```{r}
chl.timeseries <- chl.timeseries %>% 
  group_by(month) %>% 
  mutate(chla.sog.clim = mean(chla.sog, na.rm = T)) %>% 
  ungroup() %>% 
  # Calculate the anomaly time series
  mutate(chla.sog.anom = chla.sog - chla.sog.clim)

# Plot anomaly as bar plots relative to zero
ggplot(data = distinct(chl.timeseries, year.month, .keep_all = T ),
       aes(x = year.month, y = chla.sog.anom)) +
  geom_bar(stat = "identity")
```
